requirements.txt
aiogram==2.24
SQLAlchemy==1.4.47
python-dotenv==1.0.0